Fixing a bug in sequence preprocessing
When cuda is available, and a sequence of length = 1 is loaded, it is left on the cpu and not copied to the gpu. That creates an error: 
`RuntimeError: Expected all tensors to be on the same device, but found at least two devices, cuda:0 and cpu! (when checking arugment for argument tensors in method wrapper__cat)`